add pipeline module and move QuantumClassifierWithRiemannianPipeline to it
